# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Java Backend Challenge API",
    "version": "0.1.0",
    "description": "Sports training scheduling and registration API. All eight operations are implemented. Coach names retain the public name field and normalize ASCII whitespace. Timestamps require an explicit offset, are truncated to microseconds and return UTC. Errors use application/json with stable codes and optional fieldErrors. See README.md for setup and a complete HTTP walkthrough, and documentation/ADR-001-coach-name-expansion.md for database compatibility and rollback decisions."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local service; runtime Swagger documents implemented endpoints."
    }
  ],
...
  "paths": {
...
    "/api/participants": {
      "post": {
        "operationId": "createParticipant",
        "summary": "Create a participant",
        "tags": [
          "Participants"
        ],
        "description": "IDs identify participants; the same email may occur on multiple records.",
        "responses": {
          "201": {
            "description": "Resource created; body includes its ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Participant"
                },
                "example": {
                  "id": "22222222-2222-4222-8222-222222222222",
                  "name": "Sam Lee",
                  "email": "sam@example.com"
                }
              }
            }
          },
          "400": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "invalidFields": {
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "Request contains invalid fields.",
                      "instance": "/api/participants",
                      "code": "INVALID_REQUEST",
                      "fieldErrors": [
                        {
                          "field": "email",
                          "message": "must be a valid email"
                        }
                      ]
                    }
                  },
                  "malformedJson": {
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "Request body must contain valid JSON.",
                      "instance": "/api/participants",
                      "code": "INVALID_REQUEST"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "type": "about:blank",
                      "title": "Internal Server Error",
                      "status": 500,
                      "detail": "An unexpected error occurred.",
                      "instance": "/api/participants",
                      "code": "INTERNAL_ERROR"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParticipant"
              },
              "example": {
                "name": "Sam Lee",
                "email": "sam@example.com"
              }
            }
          }
        }
      }
    },
...
    "/api/sessions/{sessionId}": {
      "parameters": [
        {
          "name": "sessionId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "delete": {
        "operationId": "deleteSession",
        "summary": "Delete an empty session",
        "tags": [
          "Sessions"
        ],
        "description": "Check session existence first. Block while any registrations exist with SESSION_HAS_REGISTRATIONS. In one transaction, execute a delete by ID and inspect its affected row count; zero rows return SESSION_NOT_FOUND. Translate only a violation of fk_registrations_session into SESSION_HAS_REGISTRATIONS if a registration appears after the check. After cancellation of all registrations, deletion succeeds. Repeated deletion returns 404. Keep coaches and participants.",
        "responses": {
          "204": {
            "description": "Deleted; no response body."
          },
          "400": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "invalidUuid": {
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "Bad Request",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111",
                      "code": "INVALID_REQUEST"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "SESSION_NOT_FOUND": {
                    "value": {
                      "type": "about:blank",
                      "title": "Not Found",
                      "status": 404,
                      "detail": "Session was not found.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111",
                      "code": "SESSION_NOT_FOUND"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "SESSION_HAS_REGISTRATIONS": {
                    "value": {
                      "type": "about:blank",
                      "title": "Conflict",
                      "status": 409,
                      "detail": "Cancel all registrations before deleting this session.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111",
                      "code": "SESSION_HAS_REGISTRATIONS"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "type": "about:blank",
                      "title": "Internal Server Error",
                      "status": 500,
                      "detail": "An unexpected error occurred.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111",
                      "code": "INTERNAL_ERROR"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sessionId}/registrations": {
      "parameters": [
        {
          "name": "sessionId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "createRegistration",
        "summary": "Register a participant",
        "tags": [
          "Registrations"
        ],
        "description": "Validate input, then check session existence, participant existence, duplicate and capacity in one transaction. Duplicate takes precedence over full. The last available slot succeeds. Flush the insert before returning the generated registration ID; translate only a violation of uq_registrations_session_participant to DUPLICATE_REGISTRATION. Other unexpected persistence failures remain sanitized INTERNAL_ERROR responses. Transactional baseline capacity checking does not guarantee capacity under concurrent requests; optional JBC-017 adds that guarantee.",
        "responses": {
          "201": {
            "description": "Resource created; body includes its ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Registration"
                },
                "example": {
                  "id": "44444444-4444-4444-8444-444444444444",
                  "sessionId": "33333333-3333-4333-8333-333333333333",
                  "participantId": "22222222-2222-4222-8222-222222222222"
                }
              }
            }
          },
          "400": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "missingParticipant": {
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "Request contains invalid fields.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/registrations",
                      "code": "INVALID_REQUEST",
                      "fieldErrors": [
                        {
                          "field": "participantId",
                          "message": "must not be null"
                        }
                      ]
                    }
                  },
                  "malformedJsonOrParticipantUuid": {
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "Request body must contain valid JSON.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/registrations",
                      "code": "INVALID_REQUEST"
                    }
                  },
                  "invalidSessionUuid": {
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "Bad Request",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/registrations",
                      "code": "INVALID_REQUEST"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "SESSION_NOT_FOUND": {
                    "value": {
                      "type": "about:blank",
                      "title": "Not Found",
                      "status": 404,
                      "detail": "Session was not found.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/registrations",
                      "code": "SESSION_NOT_FOUND"
                    }
                  },
                  "PARTICIPANT_NOT_FOUND": {
                    "value": {
                      "type": "about:blank",
                      "title": "Not Found",
                      "status": 404,
                      "detail": "Participant was not found.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/registrations",
                      "code": "PARTICIPANT_NOT_FOUND"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "DUPLICATE_REGISTRATION": {
                    "value": {
                      "type": "about:blank",
                      "title": "Conflict",
                      "status": 409,
                      "detail": "Participant is already registered for this session.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/registrations",
                      "code": "DUPLICATE_REGISTRATION"
                    }
                  },
                  "SESSION_FULL": {
                    "value": {
                      "type": "about:blank",
                      "title": "Conflict",
                      "status": 409,
                      "detail": "Session has reached its capacity.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/registrations",
                      "code": "SESSION_FULL"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "type": "about:blank",
                      "title": "Internal Server Error",
                      "status": 500,
                      "detail": "An unexpected error occurred.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/registrations",
                      "code": "INTERNAL_ERROR"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegistration"
              },
              "example": {
                "participantId": "22222222-2222-4222-8222-222222222222"
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sessionId}/participants": {
      "parameters": [
        {
          "name": "sessionId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "listSessionParticipants",
        "summary": "List registered participants",
        "tags": [
          "Registrations"
        ],
        "description": "Include participant id, name, email and registrationId for discoverable cancellation. Order by participant id. An existing empty session returns empty content and zero totals; a missing session returns 404. One read-only transaction checks existence then fetches registrations with participants through an EntityGraph, avoiding per-participant queries. Returns {content,page,size,totalElements,totalPages}. Pages beyond the last return empty content with matching totals. Invalid pagination returns 400 before resource lookup.",
        "responses": {
          "200": {
            "description": "Successful query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredParticipantPage"
                },
                "example": {
                  "content": [
                    {
                      "id": "22222222-2222-4222-8222-222222222222",
                      "name": "Sam Lee",
                      "email": "sam@example.com",
                      "registrationId": "44444444-4444-4444-8444-444444444444"
                    }
                  ],
                  "page": 0,
                  "size": 20,
                  "totalElements": 1,
                  "totalPages": 1
                }
              }
            }
          },
          "400": {
            "description": "Invalid filters, identifiers or pagination parameters.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "invalidUuid": {
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "Bad Request",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/participants",
                      "code": "INVALID_REQUEST"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "SESSION_NOT_FOUND": {
                    "value": {
                      "type": "about:blank",
                      "title": "Not Found",
                      "status": 404,
                      "detail": "Session was not found.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/participants",
                      "code": "SESSION_NOT_FOUND"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "type": "about:blank",
                      "title": "Internal Server Error",
                      "status": 500,
                      "detail": "An unexpected error occurred.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/participants",
                      "code": "INTERNAL_ERROR"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Zero-based page. Omitted or empty values use 0.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "description": "Page size. Omitted or empty values use 20; maximum 100.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
    "/api/sessions/{sessionId}/registrations/{registrationId}": {
      "parameters": [
        {
          "name": "sessionId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "registrationId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "delete": {
        "operationId": "cancelRegistration",
        "summary": "Cancel a registration",
        "tags": [
          "Registrations"
        ],
        "description": "Check session existence, then execute a single delete scoped by sessionId and registrationId in one transaction. Missing registration or one belonging to another session returns 404. Delete only the relationship, freeing capacity; registering again creates a new ID. Repeated cancellation returns 404. Preserve participants and other registrations.",
        "responses": {
          "204": {
            "description": "Deleted; no response body."
          },
          "400": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "invalidUuid": {
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "Bad Request",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/registrations/11111111-1111-4111-8111-111111111111",
                      "code": "INVALID_REQUEST"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "SESSION_NOT_FOUND": {
                    "value": {
                      "type": "about:blank",
                      "title": "Not Found",
                      "status": 404,
                      "detail": "Session was not found.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/registrations/11111111-1111-4111-8111-111111111111",
                      "code": "SESSION_NOT_FOUND"
                    }
                  },
                  "REGISTRATION_NOT_FOUND": {
                    "value": {
                      "type": "about:blank",
                      "title": "Not Found",
                      "status": 404,
                      "detail": "Registration was not found under this session.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/registrations/11111111-1111-4111-8111-111111111111",
                      "code": "REGISTRATION_NOT_FOUND"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "See error codes and examples.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "type": "about:blank",
                      "title": "Internal Server Error",
                      "status": 500,
                      "detail": "An unexpected error occurred.",
                      "instance": "/api/sessions/11111111-1111-4111-8111-111111111111/registrations/11111111-1111-4111-8111-111111111111",
                      "code": "INTERNAL_ERROR"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateParticipant": {
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "description": "Must contain a non-whitespace character.",
            "example": "Sam Lee"
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1,
            "pattern": "\\S",
            "example": "sam@example.com"
          }
        }
      },
      "Participant": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "22222222-2222-4222-8222-222222222222"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "description": "Must contain a non-whitespace character.",
            "example": "Sam Lee"
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1,
            "pattern": "\\S",
            "example": "sam@example.com"
          }
        }
      },
...
      "CreateRegistration": {
        "type": "object",
        "required": [
          "participantId"
        ],
        "properties": {
          "participantId": {
            "type": "string",
            "format": "uuid",
            "example": "22222222-2222-4222-8222-222222222222"
          }
        }
      },
      "Registration": {
        "type": "object",
        "required": [
          "id",
          "sessionId",
          "participantId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "44444444-4444-4444-8444-444444444444"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid",
            "example": "33333333-3333-4333-8333-333333333333"
          },
          "participantId": {
            "type": "string",
            "format": "uuid",
            "example": "22222222-2222-4222-8222-222222222222"
          }
        }
      },
...
      "ApiProblem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Bad Request"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Request contains invalid fields."
          },
          "instance": {
            "type": "string",
            "format": "uri-reference",
            "example": "/api/coaches"
          },
          "code": {
            "type": "string",
            "enum": [
              "INVALID_REQUEST",
              "COACH_NOT_FOUND",
              "PARTICIPANT_NOT_FOUND",
              "SESSION_NOT_FOUND",
              "REGISTRATION_NOT_FOUND",
              "COACH_OVERLAP",
              "DUPLICATE_REGISTRATION",
              "SESSION_FULL",
              "SESSION_HAS_REGISTRATIONS",
              "INTERNAL_ERROR"
            ],
            "example": "INVALID_REQUEST"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "field",
                "message"
              ],
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            },
            "description": "Present only for body field validation failures; sorted, deduplicated and omitted when empty."
          }
        }
      }
    }
  },
...
}